feat: 사이드바, 갤러리 연결
