+++ src/App.tsx
@@ -4,7 +4,7 @@ import Gallery from './pages/Gallery/Gallery';
 export default function App(): JSX.Element {
   return (
     <Routes>
-      <Route path="/gallery" element={<Gallery />} />
+      <Route path="/gallery/:id" element={<Gallery />} />
     </Routes>
   );
 }
